+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0
+
+- Remove redundant extensions on `FormzInputStatus`
+- Add `isValidated` extension on `FormzStatus`
+- Add `invalid` getter to `FormzInput`
+
 # 0.1.0
 
 - Add extensions on `FormzStatus`